Add files via upload
Download both files and simply double click the html document in your downloads folder to open the webpage. You will be able to play the game with commands. Both files have to be in the same directory in order for the web game to operate. If you are experiencing any issues, feel free to contact me via email : bdiarra@texnspex.com